making blockdss package
The blockdss package can be used for modeling the spatial distribution of a disease. Simulated risk maps, median risk map and risk uncertainty map are obtained in the end.
The package provides functions to configure a file with parameters and arguments needed to invoke dss.c.64.exe and run block simulation program.